TEIID-4141: moving the auto-generate REST war into it's own module, to keep the swagger, fasterxml dependencies clearly defined
