--- jboss-integration/kits/wildfly/modules/com/fasterxml/jackson/dataformat/yaml/main/module.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module xmlns="urn:jboss:module:1.0" name="com.fasterxml.jackson.dataformat.yaml">
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-    <resources>
-        <resource-root path="jackson-dataformat-yaml-${version.com.fasterxml.jackson.swagger}.jar" />
-    </resources>
-    
-    <dependencies>
-    	<module name="org.yaml.snakeyaml" />
-    	<module name="com.fasterxml.jackson.core.jackson-core" />
-    </dependencies>
-</module>
