--- jboss-integration/kits/wildfly/modules/com/fasterxml/jackson/datatype/main/module.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module xmlns="urn:jboss:module:1.0" name="com.fasterxml.jackson.datatype">
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-    <resources>
-        <resource-root path="jackson-datatype-joda-${version.com.fasterxml.jackson.swagger}.jar" />
-    </resources>
-    
-    <dependencies>
-        <module name="org.joda.time" /> 
-    </dependencies>
-    
-</module>
