--- jboss-integration/kits/wildfly/modules/io/swagger/main/module.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module xmlns="urn:jboss:module:1.0" name="io.swagger">
-
-	<properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-    
-    <resources>
-        <resource-root path="swagger-annotations-${version.swagger-jaxrs}.jar" />
-        <resource-root path="swagger-core-${version.swagger-jaxrs}.jar" />
-        <resource-root path="swagger-jaxrs-${version.swagger-jaxrs}.jar" />
-        <resource-root path="swagger-models-${version.swagger-jaxrs}.jar" />
-    </resources>
-    
-    <dependencies>
-        <module name="com.fasterxml.jackson.core.jackson-core" />
-        <module name="com.fasterxml.jackson.core.jackson-databind"/>
-        <module name="com.fasterxml.jackson.core.jackson-annotations" />       
-        <module name="com.fasterxml.jackson.dataformat.yaml" />
-        <module name="com.fasterxml.jackson.dataformat.xml" />
-        <module name="com.fasterxml.jackson.datatype" />
-        <module name="com.fasterxml.jackson.jaxrs.jackson-jaxrs-json-provider" />
-        <module name="com.google.guava" />  
-        <module name="org.apache.commons.lang" slot="3.3.2"/>
-        <module name="org.reflections" /> 
-        <module name="org.slf4j"/>
-        <module name="javax.ws.rs.api"/>
-        <module name="javax.servlet.api"/>
-        <module name="javax.xml.bind.api"/>
-    </dependencies>
-    
-</module>
