+++ jboss-integration/kits/wildfly/modules/org/jboss/teiid/main/module.xml
@@ -21,6 +21,7 @@
         <module name="org.jboss.teiid.api" />
         <module name="org.jboss.teiid.client" />
         <module name="org.jboss.teiid.admin" />
+        <module name="org.jboss.teiid.rest-service" />
         <module name="org.jboss.logging" />
         <module name="org.jboss.modules"/>
         <module name="org.jboss.msc" />
@@ -38,8 +39,7 @@
         <module name="javax.activation.api"/>
         <module name="org.jboss.as.clustering.jgroups"/>
         <module name="org.infinispan"/>
-        <module name="asm.asm"/>
-        <module name="org.infinispan"/>
+        
         <module name="org.jboss.marshalling"/>
         <module name="org.jboss.marshalling.river"/> 
         <module name="org.jboss.as.ee"/>
@@ -47,10 +47,6 @@
         <module name="org.wildfly.clustering.jgroups.api"/>
         <module name="org.wildfly.clustering.jgroups.spi"/>  
         
-        <module name="com.fasterxml.jackson.core.jackson-core"/>
-        <module name="com.fasterxml.jackson.core.jackson-databind"/>
-        <module name="com.fasterxml.jackson.core.jackson-annotations"/>
-             
         <!-- These dependencies here for ra.xml description -->
         <module name="org.jboss.as.connector"/>
         <module name="org.jboss.ironjacamar.api"/>
@@ -63,18 +59,9 @@
         <module name="org.osgeo.proj4j"/>
         <module name="org.wololo.jts2geojson"/>
         <module name="org.jaxen" />
-        <module name="org.jboss.resteasy.resteasy-jaxrs"/> <!--  REST Web Service --> 
-        <module name="org.jboss.resteasy.resteasy-multipart-provider"/> <!--  REST Web Service -->
-        <module name="javax.ws.rs.api"/> <!--  REST Web Service -->
         <module name="org.jboss.dmr" /> <!-- Admin -->        
         <module name="org.jboss.as.cli"/> <!-- Admin -->
         <module name="org.jboss.as.controller-client"/> <!-- Admin -->
         <module name="org.jboss.as.controller"/> <!-- Admin -->
-        <module name="io.swagger"/>
-	<module name="javax.servlet.api"/>
-	<module name="org.apache.commons.io"/>
-	<module name="org.jboss.resteasy.resteasy-jaxb-provider" />
-        
     </dependencies>
-
 </module>
