+++ jboss-integration/kits/wildfly/modules/org/wololo/jts2geojson/main/module.xml
@@ -8,8 +8,6 @@
     </resources>
     <dependencies>
         <module name="com.vividsolutions.jts" />
-        <module name="com.fasterxml.jackson.core.jackson-core"/>
-        <module name="com.fasterxml.jackson.core.jackson-databind"/>
-        <module name="com.fasterxml.jackson.core.jackson-annotations"/>
+        <module name="com.fasterxml.jackson.core"/>
     </dependencies>
 </module>
