+++ jboss-integration/src/main/resources/org/teiid/jboss/i18n.properties
@@ -68,7 +68,6 @@ TEIID50075=No "path" information found to load the FUNCTION model {0}; FUNCTION 
 TEIID50076=For {0}.{1} VDB, recursive delegation {2} found.
 TEIID50077=The translator type "{0}" was not found in configuration; VDB {1}.{2} will not be started until translator is available.
 TEIID50091=The URI parameters defined on procedure {0}, do not match to number of parameters defined on procedure. Correct and reploy VDB to expose this procedure through REST. 
-TEIID50092=Only result produceing procedures are allowed
 TEIID50093=Cache factory not found; Make sure the Infinispan based cache factory is correctly configured for Teiid.
 TEIID50094=Resultset cache configured without the Infinispan's Cache Container name. Check and provide <resultset-cache infinispan-container="{name}"/> in configuration.
 TEIID50095=PreparedPlan cache configured without the Infinispan's Cache Container name. Check and provide <preparedplan-cache infinispan-container="{name}"/> in configuration.
@@ -79,7 +78,7 @@ failed_to_add= Failed to deploy {0}
 TEIID50099=Invalid JNDI name "{0}" specified in VDB {1} {2}
 TEIID50100=Trigger execution result: {0}
 TEIID50103=Wrong Response returned from {0} security domain; Expecting a Kerberoes response
-TEIID50105=Invalid form parameter; No column with name {0} defined on procedure {1}
+
 security_not_correct=Security settings on {0} transport may not be correct since it is using deprecated "krb5-domain" value.
 
 # subsystem description
