--- jboss-integration/src/main/resources/rest-war/MANIFEST.MF
@@ -1,2 +1,0 @@
-Manifest-Version: 1.0
-Dependencies: org.jboss.teiid, org.jboss.teiid.client, io.swagger
