+++ olingo-common/kits/wildfly/modules/org/apache/olingo/commons/main/module.xml
@@ -14,9 +14,7 @@
         <module name="javax.servlet.api"/>
         <module name="javax.xml.stream.api"/>
         <module name="org.apache.httpcomponents"/>
-        <module name="com.fasterxml.jackson.core.jackson-core"/>
-        <module name="com.fasterxml.jackson.core.jackson-databind"/>
-        <module name="com.fasterxml.jackson.core.jackson-annotations"/>
+        <module name="com.fasterxml.jackson.core"/>
         <module name="org.slf4j"/>
         <module name="org.codehaus.woodstox"/>
         <module name="org.apache.commons.codec"/>
