+++ pom.xml
@@ -78,8 +78,7 @@
         <version.org.eclipse.persistence>2.5.0</version.org.eclipse.persistence>
         <version.org.apache.olingo>4.2.0-teiid-3</version.org.apache.olingo>
         <version.org.eclipse.jetty>9.2.7.v20150116</version.org.eclipse.jetty>
-        <version.com.fasterxml.jackson.core>2.4.2</version.com.fasterxml.jackson.core>
-	    <version.com.fasterxml.jackson.swagger>2.4.5</version.com.fasterxml.jackson.swagger>
+        <version.com.fasterxml.jackson.core>2.4.5</version.com.fasterxml.jackson.core>
 	    <version.com.fasterxml.aalto-xml>0.9.10</version.com.fasterxml.aalto-xml>
         <version.org.reflections>0.9.9</version.org.reflections>
         <version.com.google.code.findbugs>2.0.1</version.com.google.code.findbugs>
@@ -89,7 +88,8 @@
         <version.jts>1.13</version.jts>
         <version.jts2geojson>0.7.0</version.jts2geojson>
         <version.proj4j>0.1.0</version.proj4j>
-	    <version.swagger-jaxrs>1.5.2</version.swagger-jaxrs>
+        <version.swagger-parser>1.0.10</version.swagger-parser>
+	    <version.io.swagger>1.5.2</version.io.swagger>
         <version.org.apache.phoenix.phoenix-core>4.2.1</version.org.apache.phoenix.phoenix-core>
         <version.net.sf.retrotranslator>1.2.9</version.net.sf.retrotranslator>
         <version.org.apache.cxf>3.0.5</version.org.apache.cxf>
@@ -102,7 +102,9 @@
         <version.org.springframework.asm>3.1.4.RELEASE</version.org.springframework.asm>
         <version.commons-logging>1.1</version.commons-logging>
         <version.commons-configuration>1.6</version.commons-configuration>
-        <version.org.apache.ant>1.8.2</version.org.apache.ant>	    	    
+        <version.org.apache.ant>1.8.2</version.org.apache.ant>
+        <version.com.fasterxml.jackson.dataformat>2.4.5</version.com.fasterxml.jackson.dataformat>	 
+        <version.javax.validation>1.1.0.Final</version.javax.validation>   	    
 	</properties>
 	<scm>
         <connection>scm:git:git://github.com/teiid/teiid.git</connection>
@@ -565,6 +567,11 @@
 				<version>${project.version}</version>
 				<scope>test</scope>
 			</dependency>
+	        <dependency>
+	            <groupId>org.jboss.teiid</groupId>
+	            <artifactId>rest-service</artifactId>
+	            <version>${project.version}</version>
+	        </dependency>			
 			<dependency>
 				<groupId>org.jboss.teiid</groupId>
 				<artifactId>teiid-metadata</artifactId>
@@ -1023,13 +1030,14 @@
                 <groupId>org.apache.lucene</groupId>
                 <artifactId>lucene-core</artifactId>
                 <version>${version.org.apache.lucene}</version>
-            </dependency>                           
+            </dependency>
+            -->                            
             <dependency>
 		        <groupId>com.google.guava</groupId>
 		        <artifactId>guava</artifactId>
 		        <version>${version.com.google.guava}</version>
             </dependency>
-            --> 
+            
             <dependency>
                 <groupId>org.codehaus.woodstox</groupId>
                 <artifactId>woodstox-core-asl</artifactId>
@@ -1543,6 +1551,26 @@
                 <artifactId>commons-lang3</artifactId>
                 <version>${version.org.apache.commons}</version>
             </dependency>
+			<dependency>
+			    <groupId>javax.validation</groupId>
+			    <artifactId>validation-api</artifactId>
+			    <version>${version.javax.validation}</version>
+			</dependency>      
+			<dependency>
+			    <groupId>com.fasterxml.jackson.datatype</groupId>
+			    <artifactId>jackson-datatype-joda</artifactId>
+			    <version>${version.com.fasterxml.jackson.dataformat}</version>
+			</dependency>
+            <dependency>
+                <groupId>com.fasterxml.jackson.dataformat</groupId>
+                <artifactId>jackson-dataformat-yaml</artifactId>
+                <version>${version.com.fasterxml.jackson.dataformat}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.fasterxml.jackson.dataformat</groupId>
+                <artifactId>jackson-dataformat-xml</artifactId>
+                <version>${version.com.fasterxml.jackson.dataformat}</version>                
+            </dependency>						      
             <dependency>
                 <groupId>com.fasterxml.jackson.core</groupId>
                 <artifactId>jackson-core</artifactId>
@@ -1626,7 +1654,7 @@
 	       <dependency>
                 <groupId>io.swagger</groupId>
                 <artifactId>swagger-jaxrs</artifactId>
-                <version>${version.swagger-jaxrs}</version>
+                <version>${version.io.swagger}</version>
                 <exclusions>
                     <exclusion>
                         <groupId>javax.ws.rs</groupId>
@@ -1634,6 +1662,11 @@
                     </exclusion>
                 </exclusions>
             </dependency>
+            <dependency>
+                <groupId>io.swagger</groupId>
+                <artifactId>swagger-parser</artifactId>
+                <version>${version.swagger-parser}</version>
+            </dependency>             
             <dependency>
                 <groupId>org.jboss.narayana.jta</groupId>
                 <artifactId>narayana-jta</artifactId>
@@ -1654,27 +1687,34 @@
 	            <artifactId>commons-configuration</artifactId>
 	            <version>${version.commons-configuration}</version>
 	        </dependency>
+			<dependency>
+			    <groupId>org.reflections</groupId>
+			    <artifactId>reflections</artifactId>
+			    <version>${version.org.reflections}</version>
+			</dependency>
+            
         </dependencies>
 	</dependencyManagement>
 	<modules>
 		<module>common-core</module>
 		<module>api</module>
 		<module>client</module>
+		<module>admin</module>
 		<module>engine</module>
 		<module>connectors</module>
 		<module>metadata</module>
 		<module>runtime</module>
 		<module>adminshell</module>
 		<module>hibernate-dialect</module>
-		<module>eclipselink-platform</module>
-		<module>jboss-integration</module>
-		<module>jboss-security</module>
-		<module>test-integration</module>
-		<module>admin</module>
-		<module>jboss-admin</module>
+		<module>eclipselink-platform</module>		
 		<module>odata</module>
 		<module>olingo-common</module>
 		<module>olingo-patches</module>
 		<module>olingo</module>
+		<module>jboss-admin</module>
+		<module>jboss-integration</module>
+		<module>jboss-security</module>
+		<module>rest-service</module>
+		<module>test-integration</module>
 	</modules>
 </project>
