+++ rest-service/kits/wildfly/modules/com/fasterxml/jackson/dataformat/main/module.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module xmlns="urn:jboss:module:1.0" name="com.fasterxml.jackson.dataformat">
+    <properties>
+        <property name="jboss.api" value="private"/>
+    </properties>
+    <resources>
+        <resource-root path="jackson-dataformat-yaml-${version.com.fasterxml.jackson.dataformat}.jar" />
+        <resource-root path="jackson-dataformat-xml-${version.com.fasterxml.jackson.dataformat}.jar" />
+        <resource-root path="jackson-datatype-joda-${version.com.fasterxml.jackson.dataformat}.jar"/>
+    </resources>
+    
+    <dependencies>
+        <module name="com.fasterxml.jackson.core"/>
+        <module name="org.yaml.snakeyaml" />
+        <module name="org.codehaus.woodstox"/>
+    </dependencies>
+</module>
