+++ rest-service/kits/wildfly/modules/org/jboss/teiid/rest-service/main/module.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module xmlns="urn:jboss:module:1.0" name="org.jboss.teiid.rest-service">
+
+    <properties>
+        <property name="jboss.api" value="private"/>
+    </properties>
+    
+    <resources>
+        <resource-root path="rest-service-${project.version}.jar" />
+    </resources>
+    
+    <dependencies>
+        <module name="javax.api"/>
+        <module name="javax.servlet.api"/>
+        <module name="org.jboss.teiid.common-core" />
+        <module name="org.jboss.teiid.client" />
+        <module name="org.jboss.teiid.api"/>
+        <module name="org.jboss.teiid.admin"/>
+        <module name="org.jboss.teiid"/>
+        
+        <module name="com.fasterxml.jackson.core"/>
+        <module name="com.fasterxml.jackson.dataformat"/>
+        <module name="io.swagger.core" />
+        
+        <module name="org.jboss.resteasy.resteasy-jaxrs"/> <!--  REST Web Service --> 
+        <module name="org.jboss.resteasy.resteasy-multipart-provider"/> <!--  REST Web Service -->
+        <module name="org.jboss.resteasy.resteasy-jaxb-provider" />
+        <module name="javax.ws.rs.api"/> <!--  REST Web Service -->
+        <module name="org.apache.commons.io"/>
+        <module name="asm.asm"/>
+        <module name="org.jboss.dmr" /> <!-- Admin -->        
+        <module name="org.jboss.as.cli"/> <!-- Admin -->
+        <module name="org.jboss.as.controller-client"/> <!-- Admin -->
+        <module name="org.jboss.as.controller"/> <!-- Admin -->        
+    </dependencies>
+</module>
