+++ rest-service/kits/wildfly/wildfly-dist.xml
@@ -1,0 +1,88 @@
+<assembly>
+  
+ <id>wildfly-dist</id>
+  
+  <formats>
+    <format>zip</format>
+  </formats>
+  
+  <includeBaseDirectory>false</includeBaseDirectory>
+  <baseDirectory>modules</baseDirectory>
+
+ <fileSets>
+    <fileSet>
+        <directory>kits/wildfly/modules</directory>
+        <outputDirectory>${wildfly-module-root}</outputDirectory>     
+        <filtered>true</filtered> 
+        <includes>
+          <include>**/*</include>
+        </includes>   
+    </fileSet>    
+    <fileSet>
+        <directory>kits/wildfly/docs</directory>
+        <outputDirectory>docs</outputDirectory>     
+        <filtered>true</filtered> 
+        <includes>
+          <include>**/*</include>
+        </includes>   
+    </fileSet>       
+  </fileSets>
+  
+  <dependencySets>
+      <dependencySet>
+          <includes>
+              <include>com.fasterxml.jackson.core:jackson-core</include>
+              <include>com.fasterxml.jackson.core:jackson-databind</include>
+              <include>com.fasterxml.jackson.core:jackson-annotations</include>
+          </includes>
+          <outputDirectory>${wildfly-module-root}/com/fasterxml/jackson/core/main</outputDirectory>
+          <useTransitiveDependencies>false</useTransitiveDependencies>
+          <useTransitiveFiltering>false</useTransitiveFiltering>
+      </dependencySet>
+      <dependencySet>
+          <includes>
+              <include>com.fasterxml.jackson.dataformat:jackson-dataformat-yaml</include>
+              <include>com.fasterxml.jackson.dataformat:jackson-dataformat-xml</include>
+              <include>com.fasterxml.jackson.datatype:jackson-datatype-joda</include>
+          </includes>
+          <outputDirectory>${wildfly-module-root}/com/fasterxml/jackson/dataformat/main</outputDirectory>
+          <useTransitiveDependencies>false</useTransitiveDependencies>
+          <useTransitiveFiltering>false</useTransitiveFiltering>
+      </dependencySet>              
+      <dependencySet>
+          <includes>
+              <include>io.swagger:swagger-jaxrs</include>
+              <include>io.swagger:swagger-core</include>
+              <include>io.swagger:swagger-models</include>
+              <include>io.swagger:swagger-annotations</include>
+              <include>io.swagger:swagger-parser</include>
+          </includes>
+          <useTransitiveDependencies>true</useTransitiveDependencies>
+          <outputDirectory>${wildfly-module-root}/io/swagger/core/main</outputDirectory>
+          <useProjectArtifact>false</useProjectArtifact>
+      </dependencySet>
+      <dependencySet>
+          <outputDirectory>${wildfly-module-root}/org/jboss/teiid/rest-service/main</outputDirectory>
+          <includes>
+              <include>org.jboss.teiid:rest-service</include>
+          </includes>
+          <useProjectArtifact>true</useProjectArtifact>
+      </dependencySet>
+      <dependencySet>
+          <includes>
+              <include>org.reflections:reflections</include>
+          </includes>
+          <outputDirectory>${wildfly-module-root}/org/reflections/main</outputDirectory>
+          <useProjectArtifact>false</useProjectArtifact>
+      </dependencySet>
+      <dependencySet>
+	      <outputDirectory>${wildfly-module-root}/org/apache/commons/lang/3.3.2</outputDirectory>
+	      <useTransitiveDependencies>false</useTransitiveDependencies>
+	      <includes>
+	        <include>org.apache.commons:commons-lang3</include>
+	      </includes>
+	      <useProjectArtifact>false</useProjectArtifact>
+    </dependencySet>
+  </dependencySets>
+  
+</assembly>
