+++ rest-service/src/main/java/org/teiid/jboss/rest/BootstrapServlet.java
