+++ rest-service/src/main/resources/rest-war/api.html
