+++ rest-service/src/main/resources/rest-war/httpbasic.xml
