+++ rest-service/src/main/resources/rest-war/jboss-web.xml
