+++ rest-service/src/main/resources/rest-war/web.xml
