+++ rest-service/src/main/resources/swagger-ui-2.1.1.zip
