+++ rest-service/src/test/java/org/teiid/jboss/rest/TestRestWebArchiveBuilder.java
