+++ rest-service/src/test/java/org/teiid/jboss/rest/TestResteasyEnabler.java
@@ -1,0 +1,56 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * See the COPYRIGHT.txt file distributed with this work for information
+ * regarding copyright ownership.  Some portions may be licensed
+ * to Red Hat, Inc. under one or more contributor license agreements.
+ * 
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ * 
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+ * 02110-1301 USA.
+ */
+
+package org.teiid.jboss.rest;
+
+import org.jboss.as.controller.ModelController;
+import org.junit.Test;
+import org.mockito.Mockito;
+import org.teiid.adminapi.Model.Type;
+import org.teiid.adminapi.impl.ModelMetaData;
+import org.teiid.core.util.ExecutorUtils;
+import org.teiid.deployers.CompositeVDB;
+import org.teiid.deployers.ContainerLifeCycleListener;
+import org.teiid.deployers.TestCompositeVDB;
+import org.teiid.deployers.VirtualDatabaseException;
+import org.teiid.jboss.rest.ResteasyEnabler;
+import org.teiid.metadata.MetadataStore;
+
+@SuppressWarnings("nls")
+public class TestResteasyEnabler {
+
+	@Test public void testOtherModels() throws VirtualDatabaseException {
+		ResteasyEnabler resteasyEnabler = new ResteasyEnabler("x", "1", Mockito.mock(ModelController.class), ExecutorUtils.getDirectExecutor(), Mockito.mock(ContainerLifeCycleListener.class));
+		
+		MetadataStore ms = new MetadataStore();
+		
+		CompositeVDB vdb = TestCompositeVDB.createCompositeVDB(ms, "x");
+		vdb.getVDB().addProperty("{http://teiid.org/rest}auto-generate", "true");
+		ModelMetaData model = new ModelMetaData();
+		model.setName("other");
+		model.setModelType(Type.OTHER);
+		vdb.getVDB().addModel(model);
+		
+		resteasyEnabler.finishedDeployment("x", vdb);
+	}
+	
+}
